Split out input touch injection functions to new file of partial class
